menu: fix equipment inventory accounting (item dup/loss) + cursor wrap overflow
Three menu-session fixes from a correctness review:

1. (HIGH) Equipping an item duplicated it. apply_equip_outcome wrote the equip
   byte to the roster member but never reconciled world.inventory -- the
   EquipSession's swap (decrement equipped, restore swapped-out) ran only on a
   throwaway clone. So the equipped item stayed in the bag AND the previously
   equipped item was lost. Now applies the swap to world.inventory: the equipped
   id leaves the bag (removed at 0), a non-zero swapped-out id returns to it.
   Tests: the equip-writeback test now asserts the item left the bag, plus a new
   swap test (equip over an occupied slot returns the old item, removes the new).

2. (MEDIUM) The Status screen's "remove equipment" zeroed the slot without
   crediting the item back to the bag -- the unequipped item was destroyed.
   Now returns it. The existing unequip test asserts the item comes back.

3. (LOW) The menu cursor up-wrap `(cursor + count - 1) % count` is u8 arithmetic;
   with count 255 and cursor 254 the `cursor + count` overflows u8 before the
   modulo (debug panic / release misbehavior). Widened to u16.

Shop buy/sell math, gold bank routing, inn heal, and level-up XP/stat logic were
reviewed and found correct (no changes).

+++ crates/engine-vm/src/menu.rs
@@ -346,7 +346,10 @@ pub fn step<H: MenuHost + ?Sized>(host: &mut H, ctx: &mut MenuCtx, input: MenuIn
         Some(s) => {
             let count = host.screen_item_count(s).max(1);
             if input.up {
-                ctx.cursor = (ctx.cursor + count - 1) % count;
+                // Widen the wrap arithmetic: `cursor + count` can exceed 255
+                // (e.g. cursor 254, count 255) and overflow the u8 before the
+                // `% count`, panicking in debug builds.
+                ctx.cursor = ((ctx.cursor as u16 + count as u16 - 1) % count as u16) as u8;
             } else if input.down {
                 ctx.cursor = (ctx.cursor + 1) % count;
             }
